feat(logistics): add GPS tracking, dashboard stats, fix route CRUD (Features #62, #66, #67)
- Feature #62: Route delete works end-to-end with confirmation dialog

- Feature #66: GET /shipments/{id}/track returns GPS coordinates and progress

- Feature #67: GET /dashboard/stats returns fleet metrics (vehicles, shipments, routes)

- Fixed route duration type (int->float) and hours/minutes conversion in UI

+++ src/modules/logistics/models/route.py
@@ -29,7 +29,7 @@ class RouteBase(BaseModel):
     origin: LocationInfo = Field(..., description="Origin location")
     destination: LocationInfo = Field(..., description="Destination location")
     distance: float = Field(..., gt=0, description="Distance in kilometers")
-    estimatedDuration: int = Field(..., gt=0, description="Estimated duration in minutes")
+    estimatedDuration: float = Field(..., gt=0, description="Estimated duration in minutes")
     estimatedCost: Optional[float] = Field(None, ge=0, description="Estimated cost")
     isActive: bool = Field(True, description="Is route active")
